requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-auth==2.23.4
deltalake[gcs]==0.15.2
pandas==2.1.4
pyarrow==14.0.2
openpyxl==3.1.2
duckdb==0.10.2
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
slowapi==0.1.9

